+++ tests/fixtures/expected.md
@@ -1,7 +1,7 @@
 ## DataCap and CID Checker Report
  - Organization: `Some Company Inc`
  - Project: `My Project`
- - Client: `fxxxx1`
+ - Client: `f12345`
 ### Storage Provider Distribution
 | Provider  | Total Deals Made |              Location | Percentage |
 | :-------- | ---------------: | --------------------: | ---------: |
@@ -11,6 +11,8 @@
 | provider2 |         100.00 B | San Francisco, CA, US |     10.00% |
 | provider3 |         100.00 B |      Portland, OR, US |     10.00% |
 | provider4 |         100.00 B |      New York, NY, US |     10.00% |
+
+![Provider Distribution](./provider.png)
 ### Deal Data Replication
 | Number of Replicas | Total Deals Made | Percentage |
 | -----------------: | ---------------: | ---------: |
@@ -18,6 +20,8 @@
 |                  2 |         200.00 B |     20.00% |
 |                  3 |         300.00 B |     30.00% |
 |                  4 |         400.00 B |     40.00% |
+
+![Replication Distribution](./replica.png)
 ### Deal Data Shared with other Clients
 | Other Client | Total Deals Made | Unique CIDs |
 | -----------: | ---------------: | ----------: |
